extmod/moducryptolib: Add crypto functions module, impl reuses axTLS routines.
The API follows guidelines of https://www.python.org/dev/peps/pep-0272/ ,
but is optimized for the code size, with the idea that full PEP 0272
compatibility can be added with a simple Python wrapper mode.

The naming of the module follows (u)hashlib pattern.

At the bare minimum, this module is expected to provide:

* AES128, ECB (i.e. "null") mode, encrypt only

Implementation in this commit is based on axTLS routines, and implements
following:

* AES 128 and 256
* ECB and CBC modes
* encrypt and decrypt